Update automater.py
Ran some tests to determine the cause of the mismatch. Preemptively looks like the cause is that sometimes, packet loss is not reported, and therefore, does not end up getting set as an attribute. Going to run more tests to confirm, but if it's always packet loss, could be fixed by searching for the attr and setting to N/A if not found. If not, merging pd dfs could work, but would require loading an ever growing data set to merge. (might be able to circumvent be using datetime to auto generate a new, smaller file every day)